Rename source to input terminology in workflow designer
