+++ internal-packages/workflow-designer-ui/src/editor/properties-panel/text-generation-node-properties-panel/index.tsx
@@ -24,6 +24,7 @@ import {
 } from "../ui";
 import { GenerationPanel } from "./generation-panel";
 import { InputPanel } from "./input-panel";
+import { useConnectedSources } from "./inputs";
 import { KeyboardShortcuts } from "./keyboard-shortcuts";
 import {
 	AnthropicModelPanel,
@@ -32,7 +33,6 @@ import {
 	PerplexityModelPanel,
 } from "./model";
 import { PromptPanel } from "./prompt-panel";
-import { useConnectedSources } from "./sources";
 import { GitHubToolsPanel } from "./tools";
 
 export function TextGenerationNodePropertiesPanel({
