+++ internal-packages/workflow-designer-ui/src/editor/properties-panel/text-generation-node-properties-panel/inputs/index.ts
@@ -1,4 +1,4 @@
 export * from "./types";
 export * from "./utils";
-export * from "./use-connected-sources";
-export * from "./use-source-categories";
+export * from "./use-connected-inputs";
+export * from "./use-input-categories";
