+++ internal-packages/workflow-designer-ui/src/editor/properties-panel/text-generation-node-properties-panel/inputs/use-connected-inputs.ts
