+++ internal-packages/workflow-designer-ui/src/editor/properties-panel/text-generation-node-properties-panel/inputs/use-input-categories.ts
